# specs/001-create-a-todolist/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo List API
  version: 1.0.0
components:
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
        notes:
          type: string
        scheduled_at:
          type: string
          format: date-time
        deadline_at:
          type: string
          format: date-time
        priority:
          type: string
          enum: [low, mid, high, urgent]
        status:
          type: string
          enum: [undone, done, cancelled]
        order:
          type: integer
        project_id:
          type: integer
    Project:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        status:
          type: string
          enum: [undone, done, cancelled]
paths:
  /todos:
    get:
      summary: List all todos
      responses:
        '200':
          description: A list of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: Create a new todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
  /todos/{todo_id}:
    get:
      summary: Get a todo by ID
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    put:
      summary: Update a todo
      parameters:
        - name: todo_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
  /projects:
    get:
      summary: List all projects
      responses:
        '200':
          description: A list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{project_id}:
    get:
      summary: Get a project by ID
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      summary: Update a project
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'